requirements.txt
langchain
langchain-anthropic
langchain-community
langchain-core
langchain-huggingface
langchain-text-splitters
langgraph
faiss-cpu
sentence-transformers
chainlit
pypdf
tiktoken
duckduckgo-search
requests
python-dotenv
rank-bm25
black
pylint
pytest
